+++ apps/rebalancer/src/constants.ts
@@ -1,0 +1,5 @@
+import { FiatToken, getAnyFiatTokenDetails, getAnyFiatTokenDetailsMoonbeam, PENDULUM_USDC_AXL } from "@packages/shared";
+
+export const usdcTokenDetails = PENDULUM_USDC_AXL;
+export const brlaFiatTokenDetails = getAnyFiatTokenDetails(FiatToken.BRL);
+export const brlaMoonbeamTokenDetails = getAnyFiatTokenDetailsMoonbeam(FiatToken.BRL);
